refactor: 로딩, error, 데이터 없을 때 바로 처리하도록 수정
